Refactor catalog structure: enhance module documentation, reorganize factories, and move galaxy catalogs to products/
