+++ acm/catalogs/backends/__init__.py
@@ -1,4 +1,13 @@
-"""Backends for loading dark matter catalogs and generate galaxy catalogs."""
+"""
+Abstract backend interfaces for dark matter simulations.
+
+A backend is responsible for two things:
+  1. Loading a dark matter halo catalog from a simulation (get_dark_matter_catalog)
+  2. Populating it with galaxies via an HOD or similar model (make_galaxy_catalog)
+
+To implement a new backend, subclass SnapshotBackend or LightconeBackend
+and register it with @register_backend("<name>").
+"""
 
 from .base import (
     DarkMatterBackend,
