+++ acm/catalogs/factories/__init__.py
@@ -1,3 +1,9 @@
-"""Module containing catalog factories for creating galaxy catalogs."""
+"""
+Concrete catalog factories for snapshot-based pipelines.
 
+GalaxyCatalogFactory is the main entry point for users: it loads a dark matter
+backend, runs the HOD model at each requested redshift, and stores the resulting
+GalaxyCatalog instances for downstream use.
+"""
+from .base import BaseCatalogFactory
 from .snapshot import SnapshotCatalogFactory
