+++ acm/catalogs/factories/snapshot.py
@@ -7,9 +7,9 @@ from typing import override
 from cosmoprimo import Cosmology
 
 from acm.catalogs.backends import SnapshotBackend
-from acm.catalogs.base import BaseCatalogFactory
+from acm.catalogs.factories import BaseCatalogFactory
 from acm.catalogs.dataclasses import Tracer
-from acm.catalogs.galaxy_catalogs import SnapshotCatalog
+from acm.catalogs.products import SnapshotCatalog
 
 logger = logging.getLogger(__name__)
