+++ acm/catalogs/products/__init__.py
@@ -1,0 +1,8 @@
+"""
+Galaxy catalog products generated by the catalog factories.
+
+Galaxy catalogs store and provides access to per-tracer galaxy data. 
+They are the main output of the catalog factories and the primary input for downstream analysis pipelines.
+"""
+from .base import BaseGalaxyCatalog
+from .snapshot import SnapshotCatalog
